use same script
