+++ lib/getWebpackCommonConfig.js
@@ -2,12 +2,12 @@
 
 const path = require('path');
 const fs = require('fs');
-const resolveCwd = require('./resolveCwd');
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const assign = require('object-assign');
+const resolveCwd = require('./resolveCwd');
 const getBabelCommonConfig = require('./getBabelCommonConfig');
 const getTSCommonConfig = require('./getTSCommonConfig');
 const constants = require('./constants');
-const assign = require('object-assign');
 const replaceLib = require('./replaceLib');
 
 const tsConfig = getTSCommonConfig();
